phase4/frontend/requirements.txt
datasets==4.5.0
fastapi==0.128.8
groq==1.0.0
pandas==2.2.3
uvicorn==0.39.0
python-dotenv
httpx
pytest
jinja2
python-multipart
